+++ src/bus/dram.rs
@@ -3,7 +3,7 @@ use crate::{elfload, TrapCause};
 
 pub struct Dram {
     dram: Vec<u8>,
-    pub base_addr: u32,
+    pub base_addr: u64,
     size: usize,
 }
 
@@ -43,36 +43,36 @@ impl Dram {
 #[allow(clippy::identity_op)]
 impl Device for Dram {
     // is addr in device address space
-    fn in_range(&self, addr: u32) -> bool {
+    fn in_range(&self, addr: u64) -> bool {
         (self.base_addr..=self.base_addr + self.size as u32).contains(&addr)
     }
 
     // address to raw index
-    fn addr2index(&self, addr: u32) -> usize {
+    fn addr2index(&self, addr: u64) -> usize {
         (addr - self.base_addr) as usize
     }
 
     // get 1 byte
-    fn raw_byte(&self, addr: u32) -> u8 {
+    fn raw_byte(&self, addr: u64) -> u8 {
         let index = self.addr2index(addr);
         self.dram[index]
     }
 
     // store
-    fn store8(&mut self, addr: u32, data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
+    fn store8(&mut self, addr: u64, data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
         let index = self.addr2index(addr);
         self.dram[index] = (data & 0xFF) as u8;
         Ok(())
     }
 
-    fn store16(&mut self, addr: u32, data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
+    fn store16(&mut self, addr: u64, data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
         let index = self.addr2index(addr);
         self.dram[index + 1] = ((data >> 8) & 0xFF) as u8;
         self.dram[index + 0] = ((data >> 0) & 0xFF) as u8;
         Ok(())
     }
 
-    fn store32(&mut self, addr: u32, data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
+    fn store32(&mut self, addr: u64, data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
         let index = self.addr2index(addr);
         self.dram[index + 3] = ((data >> 24) & 0xFF) as u8;
         self.dram[index + 2] = ((data >> 16) & 0xFF) as u8;
@@ -81,7 +81,7 @@ impl Device for Dram {
         Ok(())
     }
 
-    fn store64(&mut self, addr: u32, data: i64) -> Result<(), (Option<u32>, TrapCause, String)> {
+    fn store64(&mut self, addr: u64, data: i64) -> Result<(), (Option<u32>, TrapCause, String)> {
         let index = self.addr2index(addr);
         self.dram[index + 7] = ((data >> 56) & 0xFF) as u8;
         self.dram[index + 6] = ((data >> 48) & 0xFF) as u8;
@@ -95,17 +95,17 @@ impl Device for Dram {
     }
 
     // load
-    fn load8(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    fn load8(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
         let index = self.addr2index(addr);
         Ok(self.dram[index] as i8 as i32 as u32)
     }
 
-    fn load16(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    fn load16(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
         let index = self.addr2index(addr);
         Ok(((self.dram[index + 1] as i16) << 8 | (self.dram[index + 0] as i16)) as i32 as u32)
     }
 
-    fn load32(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    fn load32(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
         let index = self.addr2index(addr);
         Ok((self.dram[index + 3] as u32) << 24
             | (self.dram[index + 2] as u32) << 16
@@ -113,7 +113,7 @@ impl Device for Dram {
             | (self.dram[index + 0] as u32))
     }
 
-    fn load64(&self, addr: u32) -> Result<u64, (Option<u32>, TrapCause, String)> {
+    fn load64(&self, addr: u64) -> Result<u64, (Option<u32>, TrapCause, String)> {
         let index = self.addr2index(addr);
         Ok((self.dram[index + 7] as u64) << 56
             | (self.dram[index + 6] as u64) << 48
@@ -125,12 +125,12 @@ impl Device for Dram {
             | (self.dram[index + 0] as u64))
     }
 
-    fn load_u8(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    fn load_u8(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
         let index = self.addr2index(addr);
         Ok(self.dram[index] as u32)
     }
 
-    fn load_u16(&self, addr: u32) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    fn load_u16(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
         let index = self.addr2index(addr);
         Ok(((self.dram[index + 1] as u16) << 8 | (self.dram[index + 0] as u16)) as u32)
     }
